save progress

+++ editor.py
@@ -1,34 +1,63 @@
 from jira import JIRA
-from PySide import QtGui, QtCore, QtUiTools
+from PySide import QtGui, QtCore, QtUiTools, QtXml
 from PIL import Image
+from time import gmtime, strftime
 import os
 import json
 import sys
 import warnings
 
-#TODO LIST:
-#   1: Different attachment format support.
-# X 2: Issue id case insensitive(all to captive)
-#   3: Filter file types
-# X 4: Show file details (time, uploader), also sortable
-# X 5: Now vs Then 
-
-#Qs:
-#   1: Date format?
-
 #Suppress all warning messages cuz they are annoying
 warnings.filterwarnings("ignore")
 
+class Date():
+    def __init__(self, date):
+        self.date_str = date
+        self.year = date[:4]
+        self.month = date[5:7]
+        self.day = date[8:10]
+        self.hour = date[11:13]
+        self.minute = date[14:16]
+        self.second = date[17:19]
+        self.time_components = [self.year, self.month, self.day, self.hour, self.minute, self.second]
+
+    def getDateSimplified(self):
+        return self.date_str[:self.date_str.find('.')]
+        
+    def getDateFormatted(self, format):
+        #TODO
+        pass
+    
+    def getTimeComponents(self):
+        return self.time_components
+        
+    def __gt__(self, other):
+        result = False
+        other_time_components = other.getTimeComponents()
+        for idx, component in enumerate(self.time_components):
+            other_component = other_time_components[idx]
+            if int(component) > int(other_component):
+                result = True
+                break
+        return result
+       
 class AttachmentEditor():
     def __init__(self, needs_auth=True):
-        self.default_size = (775, 385)
-        self.folded_size = (355, 385)
+        self.default_size = (1038, 385)
+        self.expanded_size = (1038, 648)
+        
         self.folded = False
+        self.show_log = False
+        self.log_index = 1
     
         self.app = QtGui.QApplication([])
         self.login_window = QtUiTools.QUiLoader().load("ui/login.ui")
         self.main_window = QtUiTools.QUiLoader().load("ui/main.ui")
+        
         self.jira = None
+        self.jira_server_title = ""
+        self.issue = None
+        self.current_issue_id = ""
         
         if os.path.exists("profile.json"):
             with open("profile.json", 'r') as f:
@@ -51,6 +80,8 @@ class AttachmentEditor():
         self.original_attachments = {} # copy of above
         self.selectedAttachment = ""
         
+        self.changed_attachments = {}
+        
         self.app.exec_()
 
     def showLoginWindow(self):
@@ -65,10 +96,13 @@ class AttachmentEditor():
         self.username = self.login_window.username.text()
         self.password = self.login_window.password.text()
         
+        #print self.server_url, self.username, self.password
+        
         success = False
         try:
             self.jira = self.connect_to_JIRA(self.server_url, auth=(self.username, self.password))
             success = True
+            self.current_user = self.jira.current_user()[:self.jira.current_user().find('%')]
         except Exception as e:
             print e
             self.displayWarning("Failed", "Login failed. Please retry.")
@@ -78,7 +112,9 @@ class AttachmentEditor():
             with open("profile.json", 'w') as output:
                 json.dump(user_info, output)
             self.displayInfo("Success", "You have logged in successfully!")
+            self.logged_in_user = ""
             self.launchMain()
+            self.jira_server_title = self.jira.server_info()["serverTitle"]
         else:
             self.server_url = ""
             self.username = ""
@@ -92,15 +128,17 @@ class AttachmentEditor():
         self.original_tree = self.main_window.originalTree
         self.original_tree.itemDoubleClicked.connect(self.showAttachment)
         self.original_tree.header().setClickable(True)
-        self.original_tree.header().sectionClicked.connect(self.sort_by_column)
+        self.original_tree.header().sectionClicked.connect(self.sortOriginalByColumn)
         
         self.current_tree = self.main_window.currentTree
         self.current_tree.itemDoubleClicked.connect(self.triggerEditCurrent)
         self.current_tree.itemChanged.connect(self.attachmentRenamed)
         self.current_tree.itemSelectionChanged.connect(self.attachmentSelected)
+        self.current_tree.header().setClickable(True)
+        self.current_tree.header().sectionClicked.connect(self.sortCurrentByColumn)
         
         self.main_window.updateButton.clicked.connect(self.updateAttachments)
-        self.main_window.hideOriginalButton.clicked.connect(self.foldOriginal)
+        self.main_window.showLogButton.clicked.connect(self.showLogPanel)
                 
     def getIssueById(self):
         msg = "Fetching data from JIRA server. Please wait..."
@@ -109,13 +147,15 @@ class AttachmentEditor():
         self.initial_loading_complete = False
         try:
             #Get issue object, its attachments and fill the data structures
-            self.issue = self.jira.issue(id)
+            jira_issue = self.jira.issue(id)
+            self.issue = jira_issue
+            self.current_issue_id = id
             attachments = self.issue.fields.attachment
             num_attachments = len(attachments)
-            
             #Reset data structures
             self.attachments_mapping = {}  
             self.original_attachments = {}
+            print "reset done"
             
             for idx, attachment in enumerate(attachments):
                 file_name = str(attachment)
@@ -123,8 +163,8 @@ class AttachmentEditor():
                 file_author = attachment.author.displayName
                 file_created_date = attachment.created
                 self.attachments_mapping[file_name] = (attachment.get(), attachment.id, file_sufix, file_author, file_created_date)
+                self.changed_attachments[file_name] = file_name
                 self.refreshProgressBar(num_attachments, idx + 1)
-                #print "getting", filename, self.attachments_mapping.keys()
             
             #Make a copy of the original attachments
             self.original_attachments = dict(self.attachments_mapping)
@@ -140,14 +180,16 @@ class AttachmentEditor():
         except Exception as e:
             warning_msg = "Issue <%s> not found! Please retry!" % str(id)
             self.displayWarning("Warning", warning_msg)
-            #Shoot error message dialog
-            self.main_window.issueId.clear()
+            self.main_window.issueId.setText(self.current_issue_id)
+            if self.issue:
+                self.initial_loading_complete = True
+            
         self.progress_window.close()
       
     def showAttachment(self, item, col):
         attachment_name = item.text(0)
         attachment_sufix = self.original_attachments[attachment_name][2]
-        attachment_path = "attachments/" + attachment_name + attachment_sufix
+        attachment_path = "attachments/" + attachment_name + attachment_sufix if attachment_sufix not in attachment_name else "attachments/" + attachment_name
         pic_sufixes = {".png": 1, ".PNG": 1, ".bmp": 1, ".BMP": 1, ".jpg": 1, ".JPG": 1, ".jpeg": 1, ".JPEG": 1}
         try:
             with open(attachment_path, 'wb') as output:
@@ -172,7 +214,10 @@ class AttachmentEditor():
     def triggerEditCurrent(self, item, col):
         if col != 0:
             return
-        self.current_tree.editItem(item, col)
+        try:
+            self.current_tree.editItem(item, col)
+        except Exception as e:
+            raise
     
     def attachmentSelected(self):
         if self.current_tree.currentItem():
@@ -182,25 +227,62 @@ class AttachmentEditor():
         #print "called", self.initial_loading_complete, self.selectedAttachment
         if self.initial_loading_complete and self.selectedAttachment != "":
             new_name = item.text(0)
+            hit = False
+            for original_name in self.changed_attachments:
+                if self.selectedAttachment == original_name:
+                    self.changed_attachments[original_name] = new_name
+                    #print "1st changing for %s" % (original_name)
+                    hit = True
+                    break
+            if not hit:
+                for original_name in self.changed_attachments:
+                    if self.selectedAttachment == self.changed_attachments[original_name]:
+                        self.changed_attachments[original_name] = new_name
+                        #print "2nd changing for %s" % (original_name)
+                        break
+           
             self.attachments_mapping[new_name] = self.attachments_mapping.pop(self.selectedAttachment)
             self.selectedAttachment = new_name
             
     def updateAttachments(self):
-        #Delete all changed
+        num_attachments_to_be_updated = 0
+        attachments_to_be_deleted = {}
         for attachment in self.original_attachments:
+            final_name = self.changed_attachments[attachment]
+            sufix = self.original_attachments[attachment][2]
+            if final_name == attachment or final_name + sufix == attachment:
+                continue
+            num_attachments_to_be_updated += 1
+            attachments_to_be_deleted[attachment] = final_name
+        
+        self.generateLog(attachments_to_be_deleted)
+        
+        if len(attachments_to_be_deleted) == 0:
+            return
+        
+        msg = "Deleting modified attachments!"
+        self.initializeProgressBar(msg)
+        index = 0
+        #Delete all changed
+        for attachment in attachments_to_be_deleted:
             id = self.original_attachments[attachment][1]
             self.jira.delete_attachment(id)
+            index += 1
+            self.refreshProgressBar(num_attachments_to_be_updated, index)
             
         #Re-upload all changed
         success = False
-        msg = "Updating attachments!"
+        msg = "Updating modified attachments!"
         self.initializeProgressBar(msg)
-        num_attachments = len(self.attachments_mapping)
+        num_attachments = len(self.changed_attachments)
         index = 0
-        for attachment in self.attachments_mapping:
-            sufix = self.attachments_mapping[attachment][2]
-            f = self.attachments_mapping[attachment][0]
-            file_name = attachment + sufix if sufix not in attachment else attachment
+        for attachment in attachments_to_be_deleted:
+            new_name = attachments_to_be_deleted[attachment]
+            attachment_info = self.attachments_mapping[new_name]
+            f = attachment_info[0]
+            sufix = attachment_info[2]
+            file_name = new_name + sufix if sufix not in new_name else new_name
+            
             #Create a local temp copy to upload
             with open(file_name, 'wb') as tmp_out:
                 tmp_out.write(f)
@@ -212,11 +294,14 @@ class AttachmentEditor():
                     self.refreshProgressBar(num_attachments, index)
                 except Exception as e:
                     print e
+                    
             #Delete local copy
             os.remove(file_name)
+            
         if success:
             self.displayInfo("Success", "Attachments have been renamed successfully.")
             self.selectedAttachment = ""
+            self.changed_attachments = {}
             self.getIssueById()
             #print "ori after = ", self.original_attachments.keys()
             #print "curr after = ", self.attachments_mapping.keys()
@@ -228,15 +313,15 @@ class AttachmentEditor():
         jiraInstance = JIRA(options, basic_auth = auth, max_retries=0)
         return jiraInstance
 
-    def foldOriginal(self):
-        if (not self.folded):
-            self.main_window.setFixedSize(self.folded_size[0], self.folded_size[1])
-            self.main_window.hideOriginalButton.setText(">>")
-            self.folded = True
+    def showLogPanel(self):
+        if self.show_log:
+            self.main_window.setFixedSize(self.default_size[0], self.default_size [1])
+            self.main_window.showLogButton.setText("SHOW LOG")
+            self.show_log = False
         else:
-            self.main_window.setFixedSize(self.default_size[0], self.default_size[1])
-            self.main_window.hideOriginalButton.setText("<<")
-            self.folded = False
+            self.main_window.setFixedSize(self.expanded_size[0], self.expanded_size [1])
+            self.main_window.showLogButton.setText("HIDE LOG")
+            self.show_log = True
  
     def refreshDisplayPanel(self, panel, content, editable=False, sort_key=None):
         while (panel.topLevelItemCount() > 0):
@@ -253,13 +338,15 @@ class AttachmentEditor():
             elif sort_key == "Author":
                 sorting_method = lambda x: x[1][3].split(' ')[0]
             elif sort_key == "Date":
-                pass
+                sorting_method = lambda x: x[1][4]
             sorted_data = (sorted(sorted_data, key=sorting_method))[::-1]
             
+        #print sorted_data
         for data in sorted_data:
             file_name = data[0]
             author = data[1][3]
-            date = data[1][4]
+            date = Date(data[1][4])
+            date = date.getDateSimplified()
             item = QtGui.QTreeWidgetItem()
             item.setText(0, file_name)
             item.setText(1, author)
@@ -271,15 +358,15 @@ class AttachmentEditor():
                 item = panel.topLevelItem(x)
                 item.setFlags(item.flags() | QtCore.Qt.ItemIsEditable)
     
-    def sort_by_column(self, col):
-        if col == 0:
-            self.refreshDisplayPanel(self.original_tree, self.original_attachments, sort_key="Name")
-        elif col == 1:
-            self.refreshDisplayPanel(self.original_tree, self.original_attachments, sort_key="Author")
-        elif col == 2:
-            pass
-        else:
-            print "ERROR sorting!"
+    def sortOriginalByColumn(self, col):
+        key_map = {0: "Name", 1: "Author", 2: "Date"}
+        key = key_map[col]
+        self.refreshDisplayPanel(self.original_tree, self.original_attachments, sort_key=key)
+    
+    def sortCurrentByColumn(self, col):
+        key_map = {0: "Name", 1: "Author", 2: "Date"}
+        key = key_map[col]
+        self.refreshDisplayPanel(self.current_tree, self.attachments_mapping, editable=True, sort_key=key)
     
     def parseJiraTime(self, time):
         pass
@@ -301,7 +388,15 @@ class AttachmentEditor():
     
     def displayInfo(self, title, msg):
         QtGui.QMessageBox.information(self.main_window, title, msg)
-        
+    
+    def generateLog(self, changed_attachments):
+        current = self.main_window.logPanel.toPlainText()
+        log = current
+        for original_name in changed_attachments:
+            # idx / original uploader / new uploader/ original name / new name
+            log += "%d | %s | %s | %s | %s | %s | %s\n" % (self.log_index, self.original_attachments[original_name][3], self.current_user, original_name, changed_attachments[original_name], self.jira_server_title, self.current_issue_id)
+            self.log_index += 1
+        self.main_window.logPanel.document().setPlainText(log)
     
 def testStuff():
     options = {'server': "https://hdc-tomcat-jira194.ubisoft.org/jira", 'verify':False}
@@ -319,7 +414,6 @@ def testStuff():
 
 def main():
     editor = AttachmentEditor()
-    #jira_instance.add_attachment(issue=issue, attachment=file, filename="test.jpg")
     
 if __name__ == "__main__":
     #testStuff()
